moved flipping out of Autos.java
